requirements.txt
Django==3.2.12
django-hosts==5.1
fabric==2.7.0
psycopg2-binary==2.9.3
django-extensions==3.1.5
django-cors-headers==3.11.0
django-redis==4.6.0
mongoengine==0.27.0
pymongo==4.3.3
premailer==3.10.0
celery==5.2.6

django-widget-tweaks==1.4.12
django-tables2==2.6.0
django-ckeditor==6.2.0
Pillow==9.0.1

djangorestframework==3.13.1
django-filter==21.1
markdown==3.3.6
django-mptt==0.14.0

drf_yasg2==1.19.4
django-debug-toolbar==3.2.4

flake8==4.0.1
factory_boy==2.12.0
ipython==8.12.2